Fixed clear command permission

+++ src/commands/clear.ts
@@ -15,8 +15,16 @@ const command = {
       flags: MessageFlags.Ephemeral,
     });
     const channelId = interaction.options.data[0]?.channel?.id;
-
+    const member = await interaction.guild?.members.fetch(interaction.user.id);
     const channel = interaction.guild?.channels.cache.get(channelId!) as TextChannel;
+    const hasModRole = member?.roles.cache.has(process.env.DISCORD_MOD_ROLE_ID!);
+
+    if (!hasModRole) {
+      await interaction.editReply({
+        content: '❌ You do not have permission to use this command.',
+      });
+      return;
+    }
 
     const messages = (await channel.messages.fetch({ limit: 100 })).filter(m => !m.pinned);
     await channel.bulkDelete(messages, true);
